# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Your API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/playlists": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Create a new playlist.",
        "description": "Create a new playlist in the database.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "JSON data for creating a new playlist",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreatePlaylistData"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Playlist created successfully.",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Invalid request data."
          }
        }
      },
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Get all playlists.",
        "description": "Retrieve details of all playlists in the database.",
        "responses": {
          "200": {
            "description": "List of playlists.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Playlist"
              }
            }
          }
        }
      }
    },
    "/playlists/{playlist_id}": {
      "put": {
        "tags": [
          "playlists"
        ],
        "summary": "Update an existing playlist by ID.",
        "description": "Update an existing playlist in the database by providing a PlaylistID.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "PlaylistID of the desired playlist."
          },
          {
            "name": "body",
            "in": "body",
            "description": "JSON data for updating an existing playlist",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePlaylistData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist updated successfully.",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "404": {
            "description": "Playlist not found."
          }
        }
      },
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Get playlist by ID.",
        "description": "Retrieve playlist details by providing a PlaylistID.",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "PlaylistID of the desired playlist."
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist data.",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "404": {
            "description": "Playlist not found."
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Delete playlist by ID.",
        "description": "Delete playlist by providing an PlaylistID.",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "PlaylistID of the undesired playlist."
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist data.",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "404": {
            "description": "Playlist not found."
          }
        }
      }
    },
    "/playlists/add_song": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Add Song to Playlist",
        "description": "Add a song to a playlist",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "JSON data for creating a new playlist",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SongPlayListAssociation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "$ref": "#/definitions/SongPlayListAssociation"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/remove_song": {
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Remove Song from Playlist",
        "description": "Remove a song from a playlist",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "JSON data for creating a new playlist",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SongPlayListAssociation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "$ref": "#/definitions/SongPlayListAssociation"
                }
              }
            }
          }
        }
      }
    },
...
    "/login": {
        "post": {
          "summary": "User login",
          "tags": [
            "auth"
          ],
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Credentials for user login",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "Username": {
                    "type": "string"
                  },
                  "Password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Session created successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            },
            "404": {
              "description": "User not found",
              "schema": {
                "type": "object",
                "properties": {
                  "error": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
    "/logout": {
        "delete": {
          "summary": "User logout",
          "tags": [
            "auth"
          ],
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT token",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Session deleted successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid token or user has no session",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
  },
    "definitions": {
...
      "CreatePlaylistData": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string"
          },
          "UserID": {
            "type": "integer"
          }
        }
      },
      "UpdatePlaylistData": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string"
          }
        }
      },
      "SongPlayListAssociation": {
        "type": "object",
        "properties": {
          "PlaylistID": {
            "type": "integer"
          },
          "SongID": {
            "type": "integer"
          }
        }
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "PlaylistID": {
                      "type": "integer"
                    },
                    "Name":{
                      "type": "string"
                    },
                    "UserID": {
                      "type": "integer"
                    },
                    "Songs": {
                      "type": "array",
                      "properties": {
                        "SongID": {
                          "type": "integer"
                        },
                        "Title": {
                          "type": "string"
                        },
                        "Upload_Date": {
                          "type": "string"
                        },
                        "Description": {
                          "type": "string"
                        },
                        "ArtistID": {
                          "type": "integer"
                        },
                        "AlbumID": {
                          "type": "integer"
                        }
                      }
                    }

        }
      },
...
    }
  }